% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{AcknowledgeRequest}
\alias{AcknowledgeRequest}
\title{AcknowledgeRequest Object}
\usage{
AcknowledgeRequest(subscription = NULL, ackId = NULL)
}
\arguments{
\item{subscription}{The subscription whose message is being acknowledged}

\item{ackId}{The acknowledgment ID for the message being acknowledged}
}
\value{
AcknowledgeRequest object
}
\description{
AcknowledgeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the Acknowledge method.
}
\seealso{
Other AcknowledgeRequest functions: \code{\link{subscriptions.acknowledge}}
}

