% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{Empty}
\alias{Empty}
\title{Empty Object}
\usage{
Empty()
}
\value{
Empty object
}
\description{
Empty Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An empty message that you can re-use to avoid defining duplicated emptymessages in your project. A typical example is to use it as argument or thereturn value of a service API. For instance:  service Foo {    rpc Bar (proto2.Empty) returns (proto2.Empty) { };  };BEGIN GOOGLE-INTERNALThe difference between this one and net/rpc/empty-message.proto is that1) The generated message here is in proto2 C++ API.2) The proto2.Empty has minimum dependencies   (no message_set or net/rpc dependencies)END GOOGLE-INTERNAL
}
