% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{Label}
\alias{Label}
\title{Label Object}
\usage{
Label(key = NULL, strValue = NULL, numValue = NULL)
}
\arguments{
\item{key}{The key of a label is a syntactically valid URL (as per RFC 1738) with}

\item{strValue}{A string value}

\item{numValue}{An integer value}
}
\value{
Label object
}
\description{
Label Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A key-value pair applied to a given object.
}
