% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ListSubscriptionsResponse}
\alias{ListSubscriptionsResponse}
\title{ListSubscriptionsResponse Object}
\usage{
ListSubscriptionsResponse(subscription = NULL, nextPageToken = NULL)
}
\arguments{
\item{subscription}{The subscriptions that match the request}

\item{nextPageToken}{If not empty, indicates that there are more subscriptions that match the}
}
\value{
ListSubscriptionsResponse object
}
\description{
ListSubscriptionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the ListSubscriptions method.
}
