% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ListTopicsResponse}
\alias{ListTopicsResponse}
\title{ListTopicsResponse Object}
\usage{
ListTopicsResponse(nextPageToken = NULL, topic = NULL)
}
\arguments{
\item{nextPageToken}{If not empty, indicates that there are more topics that match the request,}

\item{topic}{The resulting topics}
}
\value{
ListTopicsResponse object
}
\description{
ListTopicsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the ListTopics method.
}
