% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ModifyAckDeadlineRequest}
\alias{ModifyAckDeadlineRequest}
\title{ModifyAckDeadlineRequest Object}
\usage{
ModifyAckDeadlineRequest(ackIds = NULL, subscription = NULL, ackId = NULL,
  ackDeadlineSeconds = NULL)
}
\arguments{
\item{ackIds}{List of acknowledgment IDs}

\item{subscription}{Next Index: 5}

\item{ackId}{The acknowledgment ID}

\item{ackDeadlineSeconds}{The new ack deadline with respect to the time this request was sent to the}
}
\value{
ModifyAckDeadlineRequest object
}
\description{
ModifyAckDeadlineRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the ModifyAckDeadline method.
}
\seealso{
Other ModifyAckDeadlineRequest functions: \code{\link{subscriptions.modifyAckDeadline}}
}
