% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ModifyPushConfigRequest}
\alias{ModifyPushConfigRequest}
\title{ModifyPushConfigRequest Object}
\usage{
ModifyPushConfigRequest(pushConfig = NULL, subscription = NULL)
}
\arguments{
\item{pushConfig}{An empty <code>push_config</code> indicates that the Pub/Sub system should}

\item{subscription}{The name of the subscription}
}
\value{
ModifyPushConfigRequest object
}
\description{
ModifyPushConfigRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the ModifyPushConfig method.
}
\seealso{
Other ModifyPushConfigRequest functions: \code{\link{subscriptions.modifyPushConfig}}
}
