% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PublishBatchRequest}
\alias{PublishBatchRequest}
\title{PublishBatchRequest Object}
\usage{
PublishBatchRequest(topic = NULL, messages = NULL)
}
\arguments{
\item{topic}{The messages in the request will be published on this topic}

\item{messages}{The messages to publish}
}
\value{
PublishBatchRequest object
}
\description{
PublishBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the PublishBatch method.
}
\seealso{
Other PublishBatchRequest functions: \code{\link{topics.publishBatch}}
}

