% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PublishBatchResponse}
\alias{PublishBatchResponse}
\title{PublishBatchResponse Object}
\usage{
PublishBatchResponse(messageIds = NULL)
}
\arguments{
\item{messageIds}{The server-assigned ID of each published message, in the same order as}
}
\value{
PublishBatchResponse object
}
\description{
PublishBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the PublishBatch method.
}
