% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PublishRequest}
\alias{PublishRequest}
\title{PublishRequest Object}
\usage{
PublishRequest(topic = NULL, message = NULL)
}
\arguments{
\item{topic}{The message in the request will be published on this topic}

\item{message}{The message to publish}
}
\value{
PublishRequest object
}
\description{
PublishRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the Publish method.
}
\seealso{
Other PublishRequest functions: \code{\link{topics.publish}}
}

