% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\docType{package}
\name{pubsub_googleAuthR}
\alias{pubsub_googleAuthR}
\alias{pubsub_googleAuthR-package}
\title{Google Cloud Pub/Sub API
Provides reliable, many-to-many, asynchronous messaging between applications.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:22:01
filename: /Users/mark/dev/R/autoGoogleAPI/googlepubsubv1beta1a.auto/R/pubsub_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}
}

