% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PubsubMessage}
\alias{PubsubMessage}
\title{PubsubMessage Object}
\usage{
PubsubMessage(data = NULL, label = NULL, messageId = NULL,
  publishTime = NULL)
}
\arguments{
\item{data}{The message payload}

\item{label}{Optional list of labels for this message}

\item{messageId}{ID of this message assigned by the server at publication time}

\item{publishTime}{The time at which the message was published}
}
\value{
PubsubMessage object
}
\description{
PubsubMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A message data and its labels.
}

