% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PullBatchRequest}
\alias{PullBatchRequest}
\title{Google Cloud Pub/Sub API Objects 
Provides reliable, many-to-many, asynchronous messaging between applications.}
\usage{
PullBatchRequest(subscription = NULL, returnImmediately = NULL,
  maxEvents = NULL)
}
\arguments{
\item{subscription}{The subscription from which messages should be pulled}

\item{returnImmediately}{If this is specified as true the system will respond immediately even if}

\item{maxEvents}{The maximum number of PubsubEvents returned for this request}
}
\value{
PullBatchRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:07:23
filename: /Users/mark/dev/R/autoGoogleAPI/googlepubsubv1beta1a.auto/R/pubsub_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
PullBatchRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the PullBatch method.
}
\seealso{
Other PullBatchRequest functions: \code{\link{subscriptions.pullBatch}}
}
