% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PullBatchResponse}
\alias{PullBatchResponse}
\title{PullBatchResponse Object}
\usage{
PullBatchResponse(pullResponses = NULL)
}
\arguments{
\item{pullResponses}{Received Pub/Sub messages or status events}
}
\value{
PullBatchResponse object
}
\description{
PullBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the PullBatch method.
}
