% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PullRequest}
\alias{PullRequest}
\title{PullRequest Object}
\usage{
PullRequest(subscription = NULL, returnImmediately = NULL)
}
\arguments{
\item{subscription}{The subscription from which a message should be pulled}

\item{returnImmediately}{If this is specified as true the system will respond immediately even if}
}
\value{
PullRequest object
}
\description{
PullRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the Pull method.
}
\seealso{
Other PullRequest functions: \code{\link{subscriptions.pull}}
}
