% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PushConfig}
\alias{PushConfig}
\title{PushConfig Object}
\usage{
PushConfig(pushEndpoint = NULL)
}
\arguments{
\item{pushEndpoint}{A URL locating the endpoint to which messages should be pushed}
}
\value{
PushConfig object
}
\description{
PushConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for a push delivery endpoint.
}
