% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{subscriptions.create}
\alias{subscriptions.create}
\title{Creates a subscription on a given topic for a given subscriber.If the subscription already exists, returns ALREADY_EXISTS.If the corresponding topic doesn't exist, returns NOT_FOUND.If the name is not provided in the request, the server will assign a randomname for this subscription on the same project as the topic.}
\usage{
subscriptions.create(Subscription)
}
\arguments{
\item{Subscription}{The \link{Subscription} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}

Other Subscription functions: \code{\link{Subscription}}
}
