% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{subscriptions.delete}
\alias{subscriptions.delete}
\title{Deletes an existing subscription. All pending messages in the subscriptionare immediately dropped. Calls to Pull after deletion will returnNOT_FOUND.}
\usage{
subscriptions.delete(subscription)
}
\arguments{
\item{subscription}{The subscription to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}
}
