% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{subscriptions.pull}
\alias{subscriptions.pull}
\title{Pulls a single message from the server.If return_immediately is true, and no messages are available in thesubscription, this method returns FAILED_PRECONDITION. The system is freeto return an UNAVAILABLE error if no messages are available in areasonable amount of time (to reduce system load).}
\usage{
subscriptions.pull(PullRequest)
}
\arguments{
\item{PullRequest}{The \link{PullRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}

Other PullRequest functions: \code{\link{PullRequest}}
}
