% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{subscriptions.pullBatch}
\alias{subscriptions.pullBatch}
\title{Pulls messages from the server. Returns an empty list if there are nomessages available in the backlog. The system is free to return UNAVAILABLEif there are too many pull requests outstanding for the given subscription.}
\usage{
subscriptions.pullBatch(PullBatchRequest)
}
\arguments{
\item{PullBatchRequest}{The \link{PullBatchRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}

Other PullBatchRequest functions: \code{\link{PullBatchRequest}}
}
