% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{Topic}
\alias{Topic}
\title{Google Cloud Pub/Sub API Objects 
Provides reliable, many-to-many, asynchronous messaging between applications.}
\usage{
Topic(name = NULL)
}
\arguments{
\item{name}{Name of the topic}
}
\value{
Topic object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:22:01
filename: /Users/mark/dev/R/autoGoogleAPI/googlepubsubv1beta1a.auto/R/pubsub_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Topic Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A topic resource.
}
\seealso{
Other Topic functions: \code{\link{topics.create}}
}

