% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{topics.list}
\alias{topics.list}
\title{Lists matching topics.}
\usage{
topics.list(query = NULL, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{query}{A valid label query expression}

\item{maxResults}{Maximum number of topics to return}

\item{pageToken}{The value obtained in the last ListTopicsResponse for continuation}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}
}

