% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ListTopicsResponse}
\alias{ListTopicsResponse}
\title{ListTopicsResponse Object}
\usage{
ListTopicsResponse(topic = NULL, nextPageToken = NULL)
}
\arguments{
\item{topic}{The resulting topics}

\item{nextPageToken}{If not empty, indicates that there are more topics that match the request, and this value should be passed to the next ListTopicsRequest to continue}
}
\value{
ListTopicsResponse object
}
\description{
ListTopicsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the ListTopics method.
}

