% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PullBatchRequest}
\alias{PullBatchRequest}
\title{PullBatchRequest Object}
\usage{
PullBatchRequest(subscription = NULL, returnImmediately = NULL,
  maxEvents = NULL)
}
\arguments{
\item{subscription}{The subscription from which messages should be pulled}

\item{returnImmediately}{If this is specified as true the system will respond immediately even if it is not able to return a message in the Pull response}

\item{maxEvents}{The maximum number of PubsubEvents returned for this request}
}
\value{
PullBatchRequest object
}
\description{
PullBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the PullBatch method.
}
\seealso{
Other PullBatchRequest functions: \code{\link{subscriptions.pullBatch}}
}

