% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PullResponse}
\alias{PullResponse}
\title{PullResponse Object}
\usage{
PullResponse(ackId = NULL, pubsubEvent = NULL)
}
\arguments{
\item{ackId}{This ID must be used to acknowledge the received event or message}

\item{pubsubEvent}{A pubsub message or truncation event}
}
\value{
PullResponse object
}
\description{
PullResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Either a PubsubMessage or a truncation event. One of these two must be populated.
}

