% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{Subscription}
\alias{Subscription}
\title{Subscription Object}
\usage{
Subscription(name = NULL, topic = NULL, pushConfig = NULL,
  ackDeadlineSeconds = NULL)
}
\arguments{
\item{name}{Name of the subscription}

\item{topic}{The name of the topic from which this subscription is receiving messages}

\item{pushConfig}{If push delivery is used with this subscription, this field is used to configure it}

\item{ackDeadlineSeconds}{For either push or pull delivery, the value is the maximum time after a subscriber receives a message before the subscriber should acknowledge or Nack the message}
}
\value{
Subscription object
}
\description{
Subscription Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A subscription resource.
}
\seealso{
Other Subscription functions: \code{\link{subscriptions.create}}
}

