% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{subscriptions.acknowledge}
\alias{subscriptions.acknowledge}
\title{Acknowledges a particular received message: the Pub/Sub system can remove the given message from the subscription. Acknowledging a message whose Ack deadline has expired may succeed, but the message could have been already redelivered. Acknowledging a message more than once will not result in an error. This is only used for messages received via pull.}
\usage{
subscriptions.acknowledge(AcknowledgeRequest)
}
\arguments{
\item{AcknowledgeRequest}{The \link{AcknowledgeRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}

Other AcknowledgeRequest functions: \code{\link{AcknowledgeRequest}}
}

